chore(ci): widen clippy to all targets, fix surfaced lints
`just clippy` (and the CI job that now calls it) skipped test/example/bench
code (no --all-targets) and only ran --all-features on the kube facade.
Widen to two passes, with no hardcoded feature list so features added later
are linted automatically:

    cargo clippy --workspace --all-features --all-targets --exclude e2e
    cargo clippy --workspace --all-targets

e2e is excluded from the --all-features pass because it enables both
k8s-openapi `latest` and `mk8sv`, panicking the build script; the default
pass lints it and the `#[cfg(not(feature))]` paths instead.

Fixes the lints this surfaced in previously-unlinted code:
- kube-client: box ProviderToken::Oidc (large_enum_variant); the enum is
  private, so this is not a public API change
- kube-core: drop a redundant `&` in a panic! arg
- kube-runtime: drop redundant `as u64` casts; vec! for a >16KB test array
- allow(dead_code) on intentional test/example fixtures

+++ kube-runtime/benches/memory.rs
@@ -207,12 +207,12 @@ fn collect_stats(scenario: &str, results: &mut Vec<BenchMetric>) {
     results.push(BenchMetric {
         name: format!("{scenario} - total_allocated"),
         unit: "bytes",
-        value: stats.total_bytes as u64,
+        value: stats.total_bytes,
     });
     results.push(BenchMetric {
         name: format!("{scenario} - alloc_count"),
         unit: "allocations",
-        value: stats.total_blocks as u64,
+        value: stats.total_blocks,
     });
 }
